Refactor manager tests
