+++ lsproxy/src/lsp/manager/mod.rs
@@ -1,0 +1,6 @@
+pub(crate) mod manager;
+
+pub use manager::*;
+
+#[cfg(test)]
+mod language_tests;
